+++ src/app/participant/team-selection/page.tsx
@@ -1,7 +1,9 @@
+import CreateTeamPage from "@/app/participant/team-selection/CreateTeamPage";
+
 export default function TeamSelection() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      Team Selection
+    <main>
+      <CreateTeamPage />
     </main>
   );
 }
